% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{CreateCsv}
\alias{CreateCsv}
\title{Create CSV}
\usage{
CreateCsv(ediData, filePath)
}
\arguments{
\item{ediData}{The EDI data in format of data frame or ffdf containing the data to be inserted.}

\item{filePath}{path and file name where CSV is written}
}
\description{
Create CSV
}
\details{
Generate CSV with snake-case columns
}
