% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesssing.r
\name{DelDeviceProcess}
\alias{DelDeviceProcess}
\title{Pre-processing for DEVICE deleted in 2019.10}
\usage{
DelDeviceProcess(exelFilePath, sheetName, materialData = NULL,
  deviceCode, deviceName, endDateName, materialName,
  KoreanDictFile = "./inst/csv/tmt_Eng_Kor_translation_ANSI.csv")
}
\arguments{
\item{exelFilePath}{file path for the excel file.}

\item{sheetName}{A sheet name of interest.}

\item{materialData}{Prepared material data. Default is NULL.}

\item{deviceCode}{A column name for device EDI code.}

\item{deviceName}{A column name for device.}

\item{materialName}{A column name for the material of device.}

\item{KoreanDictFile}{Path for csv file containing translation between Korean and English. If you don't want to translate, please set this value as NULL}

\item{enddateName}{A column name for deleted date.}
}
\description{
Pre-processing for DEVICE deleted in 2019.10
}
\details{
Pre-processing for the Excel EDI file
}
