% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesssing.r
\name{DeviceProcess}
\alias{DeviceProcess}
\title{Pre-processing for DEVICE}
\usage{
DeviceProcess(
  exelFilePath,
  sheetName,
  materialData = NULL,
  deviceCode,
  deviceName,
  startDateName,
  materialName,
  KoreanDictFile = "./inst/csv/tmt_Eng_Kor_translation_ANSI.csv"
)
}
\arguments{
\item{exelFilePath}{file path for the excel file.}

\item{sheetName}{A sheet name of interest.}

\item{materialData}{Prepared material data. Default is NULL.}

\item{deviceCode}{A column name for device EDI code.}

\item{deviceName}{A column name for device.}

\item{materialName}{A column name for the material of device.}

\item{KoreanDictFile}{Path for csv file containing translation between Korean and English. If you don't want to translate, please set this value as NULL}

\item{dateName}{A column name for start date.}
}
\description{
Pre-processing for DEVICE
}
\details{
Pre-processing for the Excel EDI file
}
