% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.r
\name{ExtractNew}
\alias{ExtractNew}
\title{creating Existing Data Table}
\usage{
ExtractNew()
}
\description{
creating Existing Data Table
}
\details{
importing Existing Data Tables from DB

Extract new concepts..
}
