% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{GenerateEdiVocaTable}
\alias{GenerateEdiVocaTable}
\title{Upload masterDf to DB}
\usage{
GenerateEdiVocaTable(
  ediData,
  connectionDetails,
  vocabularyDatabaseSchema = connectionDetails$schema,
  tableName,
  useMppBulkLoadS = FALSE
)
}
\arguments{
\item{ediData}{The EDI data in format of data frame to be inserted.}

\item{connectionDetails}{connectionDetails produced by "DatabaseConnector::createconnectionDetails"}

\item{vocabularyDatabaseSchema}{The name of the DB schema where the data frame should be inserted.}

\item{tableName}{The name of the table where the data should be inserted.}

\item{useMppBulkLoad}{If using Redshift or PDW, use more performant bulk loading techniques. Setting the system environment variable "USE_MPP_BULK_LOAD" to TRUE is another way to enable this mode. Please note, Redshift requires valid S3 credentials; PDW requires valid DWLoader installation. This can only be used for permanent tables, and cannot be used to append to an existing table. Default is FALSE :D}
}
\description{
Upload masterDf to DB
}
\details{
Upload masterDf to DB
}
