% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.r
\name{NewDelDeviceProcess}
\alias{NewDelDeviceProcess}
\title{Pre-processing for DEVICE deleted in 2019.10}
\usage{
NewDelDeviceProcess(exelFilePath, sheetName, materialData = NULL,
  deviceCode, deviceName, endDateName, materialName, delMonth)
}
\arguments{
\item{exelFilePath}{file path for the excel file.}

\item{sheetName}{A sheet name of interest.}

\item{materialData}{Prepared material data. Default is NULL.}

\item{deviceCode}{A column name for device EDI code.}

\item{deviceName}{A column name for device.}

\item{materialName}{A column name for the material of device.}

\item{enddateName}{A column name for deleted date.}

\item{KoreanDictFile}{Path for csv file containing translation between Korean and English. If you don't want to translate, please set this value as NULL}
}
\description{
Pre-processing for DEVICE deleted in 2019.10
}
\details{
Pre-processing for the Excel EDI file
}
