% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.r
\name{NewEdiUpdate}
\alias{NewEdiUpdate}
\title{EDI Update}
\usage{
NewEdiUpdate(
  ediData,
  startDate,
  domainIds,
  existingVocaTable,
  connectionDetails,
  useMppBulkLoadS = FALSE
)
}
\arguments{
\item{ediData}{Restructured EDI data(SUGA, Drug, Device)}

\item{startDate}{Start date of concepts}

\item{domainIds}{Domain id each concept has}

\item{existingVocaTable}{Table of existing data on DB}

\item{connectionDetails}{Table of existing data on DB}
}
\description{
EDI Update
}
\details{
EDI Update..
}
