% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesssing.r
\name{SugaProcess}
\alias{SugaProcess}
\title{Pre-processing for SUGA}
\usage{
SugaProcess(exelFilePath, sheetName, sugaData = NULL, sugaCode,
  KoreanName, EnglishName, startDateName, sanjungName,
  KoreanDictFile = "./inst/csv/suga_Eng_Kor_translation_ANSI.csv")
}
\arguments{
\item{exelFilePath}{file path for the excel file.}

\item{sheetName}{A sheet name of interest.}

\item{sugaData}{Prepared material data. Default is NULL.}

\item{sugaCode}{A column name for device EDI code.}

\item{KoreanName}{A column name for device.}

\item{EnglishName}{A column name for device.}

\item{startDateName}{A column name for start date.}

\item{sanjungName}{A column name for the material of device.}

\item{KoreanDictFile}{Path for csv file containing translation between Korean and English. If you don't want to translate, please set this value as NULL}
}
\description{
Pre-processing for SUGA
}
\details{
Pre-processing for the Excel EDI file
}
