% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesssing.r
\name{UploadProcess}
\alias{UploadProcess}
\title{Upload masterDf to DB}
\usage{
UploadProcess(dbmsName, userName, passwordName, serverName, schemaName)
}
\arguments{
\item{dbmsName}{DB to upload}

\item{userName}{user ID}

\item{passwordName}{user password}

\item{serverName}{server to upload}

\item{schemaName}{schema}
}
\description{
Upload masterDf to DB
}
\details{
Upload masterDf to DB
}
